+++ app/src/main/java/com/bytemedrive/folder/Folder.kt
@@ -7,17 +7,4 @@ data class Folder(
     val name: String,
     val starred: Boolean = false,
     val parent: UUID?,
-) {
-    companion object {
-        fun findAllFoldersRecursively(folderId: UUID, allFolders: List<Folder>): List<Folder> {
-            val objectsInFolder = allFolders.filter { it.parent == folderId }
-
-            val objectsInSubFolders = mutableListOf<Folder>()
-            for (subfolder in objectsInFolder) {
-                objectsInSubFolders.addAll(findAllFoldersRecursively(subfolder.id, allFolders))
-            }
-
-            return objectsInFolder + objectsInSubFolders
-        }
-    }
-}
+)
